requirements.txt
Django==5.1.1
celery==5.4.0
djangorestframework==3.15.2
psycopg==3.2.2
requests==2.32.3
django-extensions==3.2.3
uWSGI==2.0.26
django-environ==0.11.2
envkey==2.4.2
setuptools==75.1.0
newrelic==9.13.0
sentry-sdk==2.14.0
slack_bolt==1.20.1
